requirements.txt
opencv-python
face-recognition
flask
numpy
Pillow